+++ Changelog.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `VK_ANDROID_external_memory_android_hardware_buffer` device extension (#769)
 - Added `VK_AMD_buffer_marker` device extension (#772)
 - Added `VK_AMD_shader_info` device extension (#773)
+- Added `VK_AMDX_shader_enqueue` device extension (#776)
 
 ### Changed